restructure wasmi_collections crate
Move all definitions into an arena submodule.
