+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 members = [
-    "crates/arena",
+    "crates/collections",
     "crates/cli",
     "crates/core",
     "crates/wasmi",
@@ -26,7 +26,7 @@ exclude = ["tests"]
 wasmi = { version = "0.32.0-beta.12", path = "crates/wasmi", default-features = false }
 wasmi_wasi = { version = "0.32.0-beta.12", path = "crates/wasi", default-features = false }
 wasmi_core = { version = "0.19.0", path = "crates/core", default-features = false }
-wasmi_arena = { version = "0.5.0", path = "crates/arena", default-features = false }
+wasmi_collections = { version = "0.1.0", path = "crates/collections", default-features = false }
 num-traits = { version = "0.2.8", default-features = false }
 
 [profile.bench]
