+++ crates/collections/Cargo.toml
@@ -1,9 +1,9 @@
 [package]
-name = "wasmi_arena"
-version = "0.5.0"
+name = "wasmi_collections"
+version = "0.1.0"
 rust-version.workspace = true
-documentation = "https://docs.rs/wasmi_arena/"
-description = "Entity arena data structures for wasmi"
+documentation = "https://docs.rs/wasmi_collections/"
+description = "Specialized data structures for the Wasmi interpreter"
 authors.workspace = true
 repository.workspace = true
 edition.workspace = true
