+++ crates/collections/src/arena/dedup.rs
