+++ crates/wasmi/Cargo.toml
@@ -21,7 +21,7 @@ exclude = [
 [dependencies]
 wasmparser = { version = "0.100.2", package = "wasmparser-nostd", default-features = false }
 wasmi_core = { workspace = true }
-wasmi_arena = { workspace = true }
+wasmi_collections = { workspace = true }
 spin = { version = "0.9", default-features = false, features = [
     "mutex",
     "spin_mutex",
@@ -44,7 +44,7 @@ criterion = { version = "0.5", default-features = false }
 default = ["std"]
 std = [
     "wasmi_core/std",
-    "wasmi_arena/std",
+    "wasmi_collections/std",
     "wasmparser/std",
     "spin/std",
     "num-traits/std",
