+++ crates/wasmi/src/engine/func_types.rs
@@ -1,6 +1,6 @@
 use super::{EngineIdx, Guarded};
 use crate::FuncType;
-use wasmi_arena::{ArenaIndex, DedupArena, GuardedEntity};
+use wasmi_collections::arena::{ArenaIndex, DedupArena, GuardedEntity};
 
 /// A raw index to a function signature entity.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
