+++ crates/wasmi/src/global.rs
@@ -1,7 +1,7 @@
 use super::{AsContext, AsContextMut, Stored};
 use crate::{core::ValueType, value::WithType, Value};
 use core::{fmt, fmt::Display, ptr::NonNull};
-use wasmi_arena::ArenaIndex;
+use wasmi_collections::arena::ArenaIndex;
 use wasmi_core::UntypedValue;
 
 /// A raw index to a global variable entity.
