+++ crates/wasmi/src/memory/data.rs
@@ -1,6 +1,6 @@
 use crate::{module, store::Stored, AsContextMut};
 use std::sync::Arc;
-use wasmi_arena::ArenaIndex;
+use wasmi_collections::arena::ArenaIndex;
 
 /// A raw index to a data segment entity.
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord)]
