+++ crates/wasmi/src/memory/mod.rs
@@ -15,7 +15,7 @@ use crate::{
     error::EntityGrowError,
     store::{Fuel, ResourceLimiterRef},
 };
-use wasmi_arena::ArenaIndex;
+use wasmi_collections::arena::ArenaIndex;
 use wasmi_core::{Pages, TrapCode};
 
 /// A raw index to a linear memory entity.
